chore : move RunResult component into ResultDisplay view
