+++ src/components/index.ts
@@ -7,4 +7,4 @@ export { Tab } from './Tab'
 export { Tag } from './Tag'
 export { Tooltip } from './Tooltip'
 export { Accordion } from './Accordion'
-export { TestResult } from './TestResult'
+export { TestResult } from '../views/Run/internal/ResultDisplay/TestResult'
