+++ src/views/Run/internal/ResultDisplay/TestResult/index.ts
@@ -1,0 +1,1 @@
+export { TestResult } from './TestResult'
