+++ src/views/Run/internal/ResultDisplay/TestResult/internal/TestResultContent/internal/TestResultContent.helpers.ts
