+++ src/views/Run/internal/ResultDisplay/TestResult/internal/TestResultTitle/internal/test-result-title.module.scss
@@ -1,5 +1,5 @@
 @use 'sass:map';
-@use '../../../../../style/variables/colors.scss';
+@use '../../../../../../../../style/variables/colors.scss';
 
 .test-result-title {
   &__status {
